fix(ingestion): stop requiring ephemeral redis url
