+++ .env.example
@@ -10,6 +10,7 @@ RABBITMQ_PASSWORD="change_me"
 
 # ─── Redis ───────────────────────────────────────────────────────────────────
 # Redis 1 — ephemeral (rate limits, pipeline cache, pub/sub). No persistence.
+# Not required by ingestion; used by config-service and planned pipeline/api-gateway services.
 REDIS_EPHEMERAL_URL="redis://localhost:6379"
 # Redis 2 — persistent AOF (idempotency keys, token blacklist, quotas)
 REDIS_PERSISTENT_URL="redis://localhost:6380"
