+++ apps/ingestion/src/config/config.module.ts
@@ -1,37 +1,12 @@
 import { Module } from '@nestjs/common'
 import { ConfigModule } from '@nestjs/config'
-import Joi from 'joi'
-
-const postgresUrlWithSchema = (expectedSchema: string) =>
-  Joi.string()
-    .required()
-    .custom((value: string, helpers) => {
-      try {
-        const url = new URL(value)
-        const schema = url.searchParams.get('schema')
-        if (schema !== expectedSchema) {
-          return helpers.error('any.invalid', {
-            message: `DATABASE_URL must include ?schema=${expectedSchema} (got: ${schema ?? 'none'})`,
-          })
-        }
-      } catch {
-        return helpers.error('any.invalid', { message: 'DATABASE_URL is not a valid URL' })
-      }
-      return value
-    })
+import { appConfigValidationSchema } from './validation.schema'
 
 @Module({
   imports: [
     ConfigModule.forRoot({
       isGlobal: true,
-      validationSchema: Joi.object({
-        NODE_ENV: Joi.string().valid('development', 'production', 'test').default('development'),
-        PORT: Joi.number().default(3001),
-        DATABASE_URL: postgresUrlWithSchema('ingestion'),
-        RABBITMQ_URL: Joi.string().required(),
-        REDIS_PERSISTENT_URL: Joi.string().required(),
-        REDIS_EPHEMERAL_URL: Joi.string().required(),
-      }),
+      validationSchema: appConfigValidationSchema,
       validationOptions: {
         abortEarly: false,
       },
